Separate service from controller
