+++ controllers/authController.js
@@ -1,9 +1,7 @@
 import { validationResult } from "express-validator";
-import bcrypt from "bcrypt";
-import JWT from "jsonwebtoken";
-import User from "../models/User.js"
+import authService from "../services/user.service.js";
 
-// Register a new user
+// Register Controller
 const register = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -13,42 +11,28 @@ const register = async (req, res) => {
   const { userName, email, password, role } = req.body;
 
   try {
-    const user = await User.findOne({ email });
-    if (user) {
-      return res.status(400).json({ message: "User already exists" });
-    }
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const newUser = new User({
+    const { token, user } = await authService.register({
       userName,
       email,
-      password: hashedPassword,
+      password,
       role,
     });
-    await newUser.save();
-
-    const token = JWT.sign(
-      { id: newUser._id, role: newUser.role },
-      process.env.JWT_SECRET,
-      { expiresIn: "1h" },
-    );
     return res.status(201).json({
       message: "User registered successfully",
       token,
-      user: {
-        id: newUser._id,
-        userName: newUser.userName,
-        email: newUser.email,
-        role: newUser.role,
-      },
+      user,
     });
   } catch (error) {
+    if (error.message === "User already exists") {
+      return res.status(400).json({ message: error.message });
+    }
     return res
       .status(500)
       .json({ message: "Server error", error: error.message });
   }
 };
 
-// Login a user
+// Login Controller
 const login = async (req, res) => {
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -56,37 +40,22 @@ const login = async (req, res) => {
   }
 
   const { email, password } = req.body;
+
   try {
-    const user = await User.findOne({ email });
-    if (!user) {
-      return res.status(400).json({ message: "Invalid Email and Password  " });
-    }
-
-    const isMatch = await bcrypt.compare(password, user.password);
-    if (!isMatch) {
-      return res.status(400).json({ message: "Invalid Email and Password" });
-    }
-
-    const token = JWT.sign(
-      { id: user._id, role: user.role },
-      process.env.JWT_SECRET,
-      { expiresIn: "1h" },
-    );
+    const { token, user } = await authService.login({ email, password });
     return res.status(200).json({
       message: "User logged in successfully",
       token,
-      user: {
-        id: user._id,
-        userName: user.userName,
-        email: user.email,
-        role: user.role,
-      },
+      user,
     });
   } catch (error) {
+    if (error.message === "Invalid Email and Password") {
+      return res.status(400).json({ message: error.message });
+    }
     return res
       .status(500)
       .json({ message: "Server error", error: error.message });
   }
 };
 
-export {register,login};
+export { register, login };
